dashboard/requirements.txt
flask>=3.0
google-auth>=2.28
google-auth-oauthlib>=1.2
google-api-python-client>=2.120
